requirements.txt
streamlit
pdfplumber
langchain
langchain-text-splitters
faiss-cpu
google-generativeai
python-dotenv
openai
sentence-transformers
matplotlib
fastapi
uvicorn
python-multipart
